fix(cloud): keep the connect-compute banner, drop the dialog rewrite
Trim this PR to just the notice: the pairing card and connect dialog
extraction reached too far for what this needed. The banner's action now
opens the workstations rail directly instead of starting a pairing flow.
